refactor components for navigation directory structure
