+++ frontend/src/components/Navigation/ProfileImage.vue
